requirements.txt
dash
flask
flask_restful
plotly_calplot
python-dotenv
requests
schedule
